Refactor event IO test for reusability
