+++ test/celeritas/io/EventIO.test.cc
@@ -6,16 +6,10 @@
 //! \file celeritas/io/EventIO.test.cc
 //---------------------------------------------------------------------------//
 #include "celeritas_config.h"
-#include "corecel/ScopedLogStorer.hh"
-#include "corecel/cont/Range.hh"
-#include "corecel/cont/Span.hh"
-#include "corecel/io/Logger.hh"
-#include "celeritas/Quantities.hh"
 #include "celeritas/io/EventReader.hh"
 #include "celeritas/io/EventWriter.hh"
-#include "celeritas/phys/ParticleParams.hh"
-#include "celeritas/phys/Primary.hh"
 
+#include "EventIOTestBase.hh"
 #include "celeritas_test.hh"
 
 #if CELERITAS_USE_HEPMC3
@@ -27,138 +21,28 @@
 #    include <HepMC3/WriterAscii.h>
 #endif
 
-using celeritas::units::MevEnergy;
-
 namespace celeritas
 {
 namespace test
 {
 //---------------------------------------------------------------------------//
 
-class EventIOTest : public Test,
+class EventIOTest : public EventIOTestBase,
                     public ::testing::WithParamInterface<char const*>
 {
   protected:
-    struct ReadAllResult
-    {
-        std::vector<int> pdg;
-        std::vector<double> energy;
-        std::vector<double> pos;
-        std::vector<double> dir;
-        std::vector<double> time;
-        std::vector<int> event;
-        std::vector<int> track;
-
-        void print_expected() const;
-    };
-
-    void SetUp() override
-    {
-        using units::ElementaryCharge;
-        using units::MevMass;
-        using units::second;
-
-        auto zero = zero_quantity();
-        constexpr auto stable = ParticleRecord::stable_decay_constant();
-
-        // Create shared standard model particle data
-        particles_ = std::make_shared<ParticleParams>(ParticleParams::Input{
-            {"proton",
-             pdg::proton(),
-             MevMass{938.27208816},
-             ElementaryCharge{1},
-             stable},
-            {"d_quark",
-             PDGNumber(1),
-             MevMass{4.7},
-             ElementaryCharge{-1.0 / 3},
-             stable},
-            {"anti_u_quark",
-             PDGNumber(-2),
-             MevMass{2.2},
-             ElementaryCharge{-2.0 / 3},
-             stable},
-            {"w_minus",
-             PDGNumber(-24),
-             MevMass{8.0379e4},
-             zero,
-             1.0 / (3.157e-25 * second)},
-            {"gamma", pdg::gamma(), zero, zero, stable},
-        });
-    }
-
-    template<class F>
-    ReadAllResult read_all(F&& read_event)
-    {
-        ReadAllResult result;
-        std::vector<Primary> primaries;
-        while (primaries = read_event(), !primaries.empty())
-        {
-            for (auto const& p : primaries)
-            {
-                result.pdg.push_back(
-                    particles_->id_to_pdg(p.particle_id).unchecked_get());
-                result.energy.push_back(p.energy.value());
-                result.pos.insert(
-                    result.pos.end(), p.position.begin(), p.position.end());
-                result.dir.insert(
-                    result.dir.end(), p.direction.begin(), p.direction.end());
-                result.time.push_back(p.time);
-                result.event.push_back(p.event_id.unchecked_get());
-                result.track.push_back(p.track_id.unchecked_get());
-            }
-        }
-        return result;
-    }
-
     void read_write(std::string const& inp_filename,
                     std::string const& out_filename)
     {
-        std::vector<Primary> primaries;
-        EventReader read_event(inp_filename, particles_);
-        EventWriter write_event(out_filename, particles_);
+        VecPrimary primaries;
+        EventReader read_event(inp_filename, this->particles());
+        EventWriter write_event(out_filename, this->particles());
         while (primaries = read_event(), !primaries.empty())
         {
             write_event(primaries);
         }
     }
-
-    std::shared_ptr<ParticleParams> particles_;
 };
-
-void EventIOTest::ReadAllResult::print_expected() const
-{
-    cout << "/*** ADD THE FOLLOWING UNIT TEST CODE ***/\n"
-            "static int const expected_pdg[] = "
-         << repr(this->pdg)
-         << ";\n"
-            "EXPECT_VEC_EQ(expected_pdg, result.pdg);\n"
-            "static double const expected_energy[] = "
-         << repr(this->energy)
-         << ";\n"
-            "EXPECT_VEC_SOFT_EQ(expected_energy, result.energy);\n"
-            "static double const expected_pos[] = "
-         << repr(this->pos)
-         << ";\n"
-            "EXPECT_VEC_SOFT_EQ(expected_pos, result.pos);\n"
-            "static double const expected_dir[] = "
-         << repr(this->dir)
-         << ";\n"
-            "EXPECT_VEC_NEAR(expected_dir, result.dir, 1e-8);\n"
-            "static double const expected_time[] = "
-         << repr(this->time)
-         << ";\n"
-            "EXPECT_VEC_SOFT_EQ(expected_time, result.time);\n"
-            "static int const expected_event[] = "
-         << repr(this->event)
-         << ";\n"
-            "EXPECT_VEC_EQ(expected_event, result.event);\n"
-            "static int const expected_track[] = "
-         << repr(this->track)
-         << ";\n"
-            "EXPECT_VEC_EQ(expected_track, result.track);\n"
-            "/*** END CODE ***/\n";
-}
 
 //---------------------------------------------------------------------------//
 // TESTS
@@ -176,10 +60,10 @@ TEST_P(EventIOTest, variety_rwr)
     this->read_write(inp_filename, out_filename);
 
     // Determine the event record format and open the file
-    EventReader read_event(out_filename, particles_);
+    EventReader read_event(out_filename, this->particles());
 
     // Read events from the event record
-    auto result = read_all(read_event);
+    auto result = this->read_all(std::ref(read_event));
     if (ext == "hepevt")
     {
         GTEST_SKIP() << "HEPEVT format sorts primaries by PDG";
@@ -230,7 +114,8 @@ TEST_P(EventIOTest, no_vertex_rwr)
     this->read_write(inp_filename, out_filename);
 
     // Read it in and check
-    auto result = this->read_all(EventReader(out_filename, particles_));
+    EventReader read_event(out_filename, this->particles());
+    auto result = this->read_all(std::ref(read_event));
 
     // clang-format off
     static int const expected_pdg[] = {22, 22, 22, 22, 22, 22, 22, 22, 22, 22,
@@ -273,95 +158,27 @@ TEST_P(EventIOTest, no_vertex_rwr)
 
 TEST_P(EventIOTest, write_read)
 {
-    std::vector<Primary> primaries = [&particles = *this->particles_] {
-        auto proton_id = particles.find(pdg::proton());
-        auto gamma_id = particles.find(pdg::gamma());
-        Primary gamma{gamma_id,
-                      MevEnergy{1.23},
-                      Real3{2, 4, 5},
-                      Real3{1, 0, 0},
-                      5.67e-9 * units::second,
-                      EventId{0},
-                      TrackId{}};
-        Primary proton{proton_id,
-                       MevEnergy{2.34},
-                       Real3{3, 5, 8},
-                       Real3{0, 1, 0},
-                       5.78e-9 * units::second,
-                       EventId{0},
-                       TrackId{}};
-        std::vector<Primary> primaries{gamma, proton, gamma, proton};
-        primaries[1].position = {-3, -4, 5};
-        primaries[3].position = primaries[2].position;
-        primaries[3].time = primaries[2].time;
-        for (auto i : range(primaries.size()))
-        {
-            primaries[i].track_id = TrackId(i * 2);
-        }
-        return primaries;
-    }();
-
     std::string const ext = this->GetParam();
     std::string filename = this->make_unique_filename(std::string{"."} + ext);
 
     // Write events
     {
-        EventWriter write_event(filename, particles_);
-        write_event(primaries);
-
-        // Add another primary and update event ID
-        primaries.push_back(primaries.back());
-        primaries.back().energy = MevEnergy{3.45};
-        for (auto& p : primaries)
-        {
-            p.event_id = EventId{1};
-        }
-        write_event(primaries);
-
-        // Write a single primary with incorrect ID
-        primaries.erase(primaries.begin() + 1, primaries.end());
-        {
-            ScopedLogStorer scoped_log_{&celeritas::self_logger()};
-            write_event(primaries);
-            static char const* const expected_log_messages[]
-                = {"Overwriting primary event IDs with 2: 1"};
-            EXPECT_VEC_EQ(expected_log_messages, scoped_log_.messages());
-            static char const* const expected_log_levels[] = {"warning"};
-            EXPECT_VEC_EQ(expected_log_levels, scoped_log_.levels());
-        }
+        EventWriter write_event(filename, this->particles());
+        this->write_test_event(std::ref(write_event));
     }
 
-    // Read events
-    auto result = this->read_all(EventReader(filename, particles_));
-
+    EventReader read_event(filename, this->particles());
     if (ext == "hepevt")
     {
+        // Read events but don't check
+        auto result = this->read_all(std::ref(read_event));
+
         GTEST_SKIP() << "HEPEVT results are nondeterministic";
     }
-
-    // clang-format off
-    static int const expected_pdg[] = {22, 2212, 22, 2212, 22, 2212, 22, 2212,
-        2212, 22};
-    static double const expected_energy[] = {1.23, 2.34, 1.23, 2.34, 1.23,
-        2.34, 1.23, 2.34, 3.45, 1.23};
-    static double const expected_pos[] = {2, 4, 5, -3, -4, 5, 2, 4, 5, 2, 4, 5,
-        2, 4, 5, -3, -4, 5, 2, 4, 5, 2, 4, 5, 2, 4, 5, 2, 4, 5};
-    static double const expected_dir[] = {1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0,
-        1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0};
-    static double const expected_time[] = {5.67e-09, 5.78e-09, 5.67e-09,
-        5.67e-09, 5.67e-09, 5.78e-09, 5.67e-09, 5.67e-09, 5.67e-09, 5.67e-09};
-    static int const expected_event[] = {0, 0, 0, 0, 1, 1, 1, 1, 1, 2};
-    static int const expected_track[] = {0, 1, 2, 3, 0, 1, 2, 3, 4, 0};
-    // clang-format on
-
-    EXPECT_VEC_EQ(expected_pdg, result.pdg);
-    EXPECT_VEC_SOFT_EQ(expected_energy, result.energy);
-    EXPECT_VEC_SOFT_EQ(expected_pos, result.pos);
-    EXPECT_VEC_SOFT_EQ(expected_dir, result.dir);
-    EXPECT_VEC_NEAR(
-        expected_time, result.time, (ext == "hepevt" ? 1e-6 : 1e-12));
-    EXPECT_VEC_EQ(expected_event, result.event);
-    EXPECT_VEC_EQ(expected_track, result.track);
+    else
+    {
+        this->read_check_test_event(std::ref(read_event));
+    }
 }
 
 INSTANTIATE_TEST_SUITE_P(EventIO,
